HIVE-29559: using cast greatly reduces the need of refactoring
