+++ ql/src/java/org/apache/hadoop/hive/ql/ddl/table/create/CreateTableAnalyzer.java
@@ -137,11 +137,6 @@ public class CreateTableAnalyzer extends CalcitePlanner {
     return analyzeAndResolveChildTree(child, plannerCtx);
   }
 
-  @Override
-  public CreateTableDesc getCreatedTableDesc() {
-    return getQB().getTableDesc();
-  }
-
   /**
    * Checks to see if given partition columns has DEFAULT or CHECK constraints (whether ENABLED or DISABLED)
    *  Or has NOT NULL constraints (only ENABLED)
